requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pillow>=10.4.0
torch>=2.1.0
torchvision>=0.16.0
python-multipart==0.0.12
numpy>=1.24.0
transformers>=4.30.0
rembg>=2.0.0
# Alternative légère pour NSFW detection
# nudenet>=2.0.0
